applied feeback
